Refactor price properties to use decimal type and update related files for consistency

+++ BookStore.Models/OrderHeader.cs
@@ -19,7 +19,7 @@ namespace Bulky.Models
 
         public DateTime ShippingDate { get; set; }  
 
-        public double OrderTotal { get; set; }
+        public decimal OrderTotal { get; set; }
 
         public string? OrderStatus { get; set; }
